+++ app/src/main/java/uz/behzoddev/restaurantpos/data/local/datasource/MenuDataSource.kt
@@ -6,6 +6,6 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodMenu
 interface MenuDataSource {
     suspend fun insertMenu(foodMenu: FoodMenu): Long
     suspend fun updateMenu(foodMenu: FoodMenu): Long
-    suspend fun deleteMenu(foodMenu: FoodMenu): Long
+    suspend fun deleteMenu(foodMenu: FoodMenu): Int
     fun fetchAllMenus(): Flow<List<FoodMenu>>
 }
